CartItem: Coomplete functionalities for calculating total amount for all items in the cart, subtotal for each item, navigating between product and cart pages, and checkout button.
CartSlice: Add two function reducers for incrementing and decrementing item quantity in the cart. Use these functions in the meantime instead of the updateQuantity since the latter does not work. Will fix this later.

ProductList: Add functionality  for displaying total quantity on the cart icon in the nav bar section.